Парсер собран в один файл
